+++ test/cypress/integration/lforms-answerConstraint-with-different-data-types.cy.js
@@ -26,7 +26,7 @@ function testAnswerConstraintNullOrOptionsOnly(valueType, params, formData, noAn
 
   describe(valueType, () => {
 
-    let desc = noAnswerConstraint ? 
+    let desc = noAnswerConstraint ?
         'should render a questionnaire with '+valueType+' and no answerConstraint' :
         'should render a questionnaire with '+valueType+' and answerConstraint="optionsOnly"'
     it(desc, function() {
@@ -37,16 +37,16 @@ function testAnswerConstraintNullOrOptionsOnly(valueType, params, formData, noAn
       cy.byId(params.itemIds.g1item1)
         .should('be.visible')
         .click();
-      cy.byCss("#searchResults li").eq(0).contains(params.itemValues.g1Answer1);
-      cy.byCss("#searchResults li").eq(1).contains(params.itemValues.g1Answer2);  
-      cy.byCss("#searchResults li").eq(2).contains(params.itemValues.g1Answer3);    
+      cy.byCss("#lhc-tools-searchResults li").eq(0).contains(params.itemValues.g1Answer1);
+      cy.byCss("#lhc-tools-searchResults li").eq(1).contains(params.itemValues.g1Answer2);
+      cy.byCss("#lhc-tools-searchResults li").eq(2).contains(params.itemValues.g1Answer3);
       // autocomplete, repeats
       cy.byId(params.itemIds.g1item2)
         .should('be.visible')
         .click();
-      cy.byCss("#searchResults li").eq(0).contains(params.itemValues.g1Answer1);
-      cy.byCss("#searchResults li").eq(1).contains(params.itemValues.g1Answer2);  
-      cy.byCss("#searchResults li").eq(2).contains(params.itemValues.g1Answer3);    
+      cy.byCss("#lhc-tools-searchResults li").eq(0).contains(params.itemValues.g1Answer1);
+      cy.byCss("#lhc-tools-searchResults li").eq(1).contains(params.itemValues.g1Answer2);
+      cy.byCss("#lhc-tools-searchResults li").eq(2).contains(params.itemValues.g1Answer3);
 
       // group 2
       // raidobutoon
@@ -63,16 +63,16 @@ function testAnswerConstraintNullOrOptionsOnly(valueType, params, formData, noAn
       cy.byId(params.itemIds.g3item1)
         .should('be.visible')
         .click();
-      cy.byCss("#searchResults li").eq(0).contains(params.itemValues.g3Answer1);
-      cy.byCss("#searchResults li").eq(1).contains(params.itemValues.g3Answer2);  
-      cy.byCss("#searchResults li").eq(2).contains(params.itemValues.g3Answer3);  
+      cy.byCss("#lhc-tools-searchResults li").eq(0).contains(params.itemValues.g3Answer1);
+      cy.byCss("#lhc-tools-searchResults li").eq(1).contains(params.itemValues.g3Answer2);
+      cy.byCss("#lhc-tools-searchResults li").eq(2).contains(params.itemValues.g3Answer3);
         // autocomplete, repeats, prefix, score
       cy.byId(params.itemIds.g3item2)
         .should('be.visible')
         .click();
-      cy.byCss("#searchResults li").eq(0).contains(params.itemValues.g3Answer1);
-      cy.byCss("#searchResults li").eq(1).contains(params.itemValues.g3Answer2);  
-      cy.byCss("#searchResults li").eq(2).contains(params.itemValues.g3Answer3);  
+      cy.byCss("#lhc-tools-searchResults li").eq(0).contains(params.itemValues.g3Answer1);
+      cy.byCss("#lhc-tools-searchResults li").eq(1).contains(params.itemValues.g3Answer2);
+      cy.byCss("#lhc-tools-searchResults li").eq(2).contains(params.itemValues.g3Answer3);
 
       // group 4
       // radiobutton, prefix, score
@@ -92,7 +92,7 @@ function testAnswerConstraintNullOrOptionsOnly(valueType, params, formData, noAn
       // autocomplete, repeats, initial
       cy.byId(params.itemIds.g5item2)
         .should('be.visible');
-     
+
       cy.byId(`item-${params.itemIds.g5item2}`)
         .byCss('span.autocomp_selected li')
         .should('have.length', 2)
@@ -104,7 +104,7 @@ function testAnswerConstraintNullOrOptionsOnly(valueType, params, formData, noAn
         .byCss('span.autocomp_selected li')
         .eq(1)
         .should('have.text', '×' + params.itemValues.g1Answer3)
-    
+
       // group 6
       // radiobutton, non-repeats, initial
       cy.byId(params.itemIds.g6item1ans2)
@@ -145,16 +145,16 @@ function testAnswerConstraintOptionsOrString(valueType, params, formData) {
       cy.byId(params.itemIds.g1item1)
         .should('be.visible')
         .click();
-      cy.byCss("#searchResults li").eq(0).contains(params.itemValues.g1Answer1);
-      cy.byCss("#searchResults li").eq(1).contains(params.itemValues.g1Answer2);  
-      cy.byCss("#searchResults li").eq(2).contains(params.itemValues.g1Answer3);    
+      cy.byCss("#lhc-tools-searchResults li").eq(0).contains(params.itemValues.g1Answer1);
+      cy.byCss("#lhc-tools-searchResults li").eq(1).contains(params.itemValues.g1Answer2);
+      cy.byCss("#lhc-tools-searchResults li").eq(2).contains(params.itemValues.g1Answer3);
       // autocomplete, repeats
       cy.byId(params.itemIds.g1item2)
         .should('be.visible')
         .click();
-      cy.byCss("#searchResults li").eq(0).contains(params.itemValues.g1Answer1);
-      cy.byCss("#searchResults li").eq(1).contains(params.itemValues.g1Answer2);  
-      cy.byCss("#searchResults li").eq(2).contains(params.itemValues.g1Answer3);    
+      cy.byCss("#lhc-tools-searchResults li").eq(0).contains(params.itemValues.g1Answer1);
+      cy.byCss("#lhc-tools-searchResults li").eq(1).contains(params.itemValues.g1Answer2);
+      cy.byCss("#lhc-tools-searchResults li").eq(2).contains(params.itemValues.g1Answer3);
 
       // group 2
       // radiobutoon
@@ -171,16 +171,16 @@ function testAnswerConstraintOptionsOrString(valueType, params, formData) {
       cy.byId(params.itemIds.g3item1)
         .should('be.visible')
         .click();
-      cy.byCss("#searchResults li").eq(0).contains(params.itemValues.g3Answer1);
-      cy.byCss("#searchResults li").eq(1).contains(params.itemValues.g3Answer2);  
-      cy.byCss("#searchResults li").eq(2).contains(params.itemValues.g3Answer3);  
+      cy.byCss("#lhc-tools-searchResults li").eq(0).contains(params.itemValues.g3Answer1);
+      cy.byCss("#lhc-tools-searchResults li").eq(1).contains(params.itemValues.g3Answer2);
+      cy.byCss("#lhc-tools-searchResults li").eq(2).contains(params.itemValues.g3Answer3);
         // autocomplete, repeats, prefix, score
       cy.byId(params.itemIds.g3item2)
         .should('be.visible')
         .click();
-      cy.byCss("#searchResults li").eq(0).contains(params.itemValues.g3Answer1);
-      cy.byCss("#searchResults li").eq(1).contains(params.itemValues.g3Answer2);  
-      cy.byCss("#searchResults li").eq(2).contains(params.itemValues.g3Answer3);  
+      cy.byCss("#lhc-tools-searchResults li").eq(0).contains(params.itemValues.g3Answer1);
+      cy.byCss("#lhc-tools-searchResults li").eq(1).contains(params.itemValues.g3Answer2);
+      cy.byCss("#lhc-tools-searchResults li").eq(2).contains(params.itemValues.g3Answer3);
 
       // group 4
       // radiobutton, prefix, score
@@ -211,16 +211,16 @@ function testAnswerConstraintOptionsOrString(valueType, params, formData) {
         .byCss('span.autocomp_selected li')
         .eq(1)
         .should('have.text', '×' + 'user typed value')
-      
+
       // group 6
       // radiobutton, non-repeats, initial
       cy.byId(params.itemIds.g6item1ans2)
         .should('be.visible')
         .should('contain',params.itemValues.g1Answer2);
-      
+
       cy.byId(`${params.itemIds.g6item1ans2} input`)
         .should('not.be.checked');
-      cy.byId(`${params.itemIds.g6item1ansOther} input`) 
+      cy.byId(`${params.itemIds.g6item1ansOther} input`)
         .should('be.checked');
       cy.byId(`${params.itemIds.g6item1ansOtherValue}`)
         .should('be.visible')
@@ -237,12 +237,12 @@ function testAnswerConstraintOptionsOrString(valueType, params, formData) {
       cy.byId(params.itemIds.g6item2ans3)
         .should('be.visible')
         .should('contain',params.itemValues.g1Answer3);
-      
+
       cy.byId(`${params.itemIds.g6item2ans3} input`)
         .should('not.be.checked');
-      cy.byId(`${params.itemIds.g6item2ansOther} input`) 
+      cy.byId(`${params.itemIds.g6item2ansOther} input`)
         .should('be.checked');
-      cy.byId(`${params.itemIds.g6item2ansOtherValue}`) 
+      cy.byId(`${params.itemIds.g6item2ansOtherValue}`)
         .should('be.visible')
         .should('have.value', "user typed value")
 
@@ -250,21 +250,21 @@ function testAnswerConstraintOptionsOrString(valueType, params, formData) {
       // radiobutton, matrix
       cy.byId(`${params.itemIds.g7item1ans2}`)
         .should('be.checked');
-      cy.byId(`${params.itemIds.g7item2ansOther}`) 
+      cy.byId(`${params.itemIds.g7item2ansOther}`)
         .should('be.checked');
       cy.byId(`${params.itemIds.g7item2ansOtherValue}`)
         .should('be.visible')
         .should('have.value', "user typed value")
-      
+
       // group 8
       // checknbox, matrix
       cy.byId(`${params.itemIds.g8item1ans2}`)
         .should('be.checked');
       cy.byId(`${params.itemIds.g8item1ans3}`)
         .should('be.checked');
-      cy.byId(`${params.itemIds.g8item2ans2}`) 
+      cy.byId(`${params.itemIds.g8item2ans2}`)
         .should('be.checked');
-      cy.byId(`${params.itemIds.g8item2ansOther}`) 
+      cy.byId(`${params.itemIds.g8item2ansOther}`)
         .should('be.checked');
       cy.byId(`${params.itemIds.g8item2ansOtherValue}`)
         .should('be.visible')
@@ -330,7 +330,7 @@ describe('answerConstaint with different types', () => {
         g3Answer3 : 'C. Answer 3 - 3'
       }
     };
-    
+
 
     let qrItemValues = {
       'valueString': {
@@ -359,9 +359,9 @@ describe('answerConstaint with different types', () => {
         g1Answer3 : {"valueCoding": {"code": "c3", "display": "Answer 3"}}
       }
     };
-    
+
     let itemIds = ((valueType, itemValues) => {
-      let coding = (valueType === "valueCoding") 
+      let coding = (valueType === "valueCoding")
 
       return {
         g1item1 : `${valueType}-group1-item1/1/1`,
@@ -425,5 +425,5 @@ describe('answerConstaint with different types', () => {
 });
 
 
-  
+